Add to_string to avoid handling version

+++ lib/src/v3.rs
@@ -628,11 +628,12 @@ impl SFrameFDE {
             .flags
             .contains(SFrameFlags::SFRAME_F_FDE_FUNC_START_PCREL)
         {
-            (self.func_start_offset as i64)
+            self.func_start_offset
                 .wrapping_add_unsigned(self.offset as u64)
                 .wrapping_add_unsigned(section.section_base) as u64
         } else {
-            (self.func_start_offset as i64).wrapping_add_unsigned(section.section_base) as u64
+            self.func_start_offset
+                .wrapping_add_unsigned(section.section_base) as u64
         }
     }
